refactor: moved configuration template classes
