+++ repositories/server-repository/src/main/kotlin/dev/redicloud/repository/server/CloudServer.kt
@@ -2,7 +2,7 @@ package dev.redicloud.repository.server
 
 import dev.redicloud.repository.service.CloudService
 import dev.redicloud.repository.service.ServiceSession
-import dev.redicloud.repository.temlate.configuration.ConfigurationTemplate
+import dev.redicloud.repository.template.configuration.ConfigurationTemplate
 import dev.redicloud.utils.service.ServiceId
 
 class CloudServer(
