+++ services/base-service/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
     compileOnly(project(":repositories:node-repository"))
     compileOnly(project(":repositories:server-repository"))
     compileOnly(project(":repositories:file-template-repository"))
+    compileOnly(project(":repositories:configuration-template-repository"))
     compileOnly(project(":repositories:server-version-repository"))
     compileOnly(project(":events"))
     compileOnly(project(":tasks"))
